Prevent homedir removal on upgrade

+++ scripts/post_uninstall.sh
@@ -4,4 +4,6 @@ set -e
 
 homedir=
 
-rm -rf ${homedir}
+if ! [[ -z "$1" ]] && ! [[ "$1" == upgrade ]] && ! [[ "$1" -eq 1 ]]; then
+  rm -rf ${homedir}
+fi
